requirements.txt
streamlit
ultralytics
pillow
numpy
opencv-python
torch
torchvision
